feat(mllp-server): check msh segment instead of pid + cleanup, move stuff to utils
refs. metriport/metriport-internal#2757
